% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Run analysis}
\usage{
runAnalysis(
  mysite,
  mask_thres,
  blur_abs,
  out_dir = "output",
  data_dir = "Input"
)
}
\arguments{
\item{mysite}{Your site acronym}

\item{mask_thres}{Obfuscation small count mask threshold (e.g. 10)}

\item{blur_abs}{Absolute max of obfuscation blurring range}

\item{out_dir}{Optional. Directory where resulting outputs are located.
Defaults to 'output'.}

\item{data_dir}{Optional. Directory where datasets are located.
Defaults to 'Input'.}
}
\value{

}
\description{
Run analysis
}
\examples{

}
